Add a link to the most recent version of CRT from the top of main branch of repository i-care.by
I think I won't publish links to particular versions of this CRT here.
Because now history of modifications of CRT is available in the
repository and there is no need to track it here.

The page should remain here for now, because I published link to this
page in public YouTube video in @i-care-by channel.